fix: update kevin bruce mason
https://baltimore.cbslocal.com/2019/03/27/body-camera-footage-released-in-police-involved-shooting-kevin-mason/

The victim had threatened to shoot officers if they didn't get off his property. In an altercation in an alley behind the house the victim pointed an object at them they believed to be a gun and fired.

No firearm was retrieved inside the house. I think "unknown" is the best fit here.